+++ caffe2/sgd/adagrad_fused.h
@@ -103,7 +103,7 @@ class SparseAdagradFusedWithSparseLengthsSumGradientOp final
         // access within range
         // gradient access within range
         CAFFE_ENFORCE_GE(
-            Input(PARAM).numel(),
+            static_cast<std::size_t>(Input(PARAM).numel()),
             block_size + offsetIdx,
             this->debug_def().input(PARAM),
             ", out of bound,  idx:",
@@ -254,7 +254,7 @@ class SparseAdagradFusedWithSparseLengthsWeightedSumGradientOp final
         // access within range
         // gradient access within range
         CAFFE_ENFORCE_GE(
-            Input(PARAM).numel(),
+            static_cast<std::size_t>(Input(PARAM).numel()),
             block_size + offsetIdx,
             this->debug_def().input(PARAM),
             ", out of bound,  idx:",
@@ -421,7 +421,7 @@ class SparseAdagradFusedWithSparseLengthsWeightedSumGradientApproxOp final
         // access within range
         // gradient access within range
         CAFFE_ENFORCE_GE(
-            Input(PARAM).numel(),
+            static_cast<std::size_t>(Input(PARAM).numel()),
             block_size + offsetIdx,
             this->debug_def().input(PARAM),
             ", out of bound,  idx:",
